hackathon_etl_v2/requirements.txt
pymongo==4.7.3
beautifulsoup4==4.12.3
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
pyyaml==6.0.1
PyPDF2==3.0.1
